requirements.txt
google-api-python-client==2.126.0
grpcio==1.62.2
protobuf==5.26.1
